[Behat] ProductVariant extraction
