+++ etc/behat/suites/domain/product.yml
@@ -11,6 +11,7 @@ default:
                 - sylius.behat.context.transform.shared_storage
                 - sylius.behat.context.transform.addressing
                 - sylius.behat.context.transform.product
+                - sylius.behat.context.transform.product_variant
                 - sylius.behat.context.transform.customer
                 - sylius.behat.context.transform.payment
                 - sylius.behat.context.transform.shipping_method
