+++ etc/behat/suites/ui/cart.yml
@@ -11,6 +11,7 @@ default:
                 - sylius.behat.context.transform.lexical
                 - sylius.behat.context.transform.shared_storage
                 - sylius.behat.context.transform.product
+                - sylius.behat.context.transform.product_variant
                 - sylius.behat.context.transform.shipping_method
                 - sylius.behat.context.transform.tax_category
